fix(LearningToCook): move the base styles
